Changes to support reuse of existing workflows (WIP)

+++ bcbio-run-workflow/src/bcbio-run-workflow.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 main() {
 
-    BCBIO_CONTAINER="record-FF2G3K00f5vGxY7V9g4Z55F2"
+    if [ "$pull_from_docker_registry" = "true"]; then
+        BCBIO_ASSETS=""
+    else
+        BCBIO_ASSETS="--assets record-FF2G3K00f5vGxY7V9g4Z55F2"
+    fi
 
     dx download "$yaml_template" -o yaml_template.yml
     dx download "$sample_spec" -o sample_spec.csv
@@ -36,15 +40,20 @@ main() {
     tar -cvzf $PNAME-generated-cwl.tgz $PNAME-workflow/main-$PNAME.cwl
     
     git clone https://github.com/dnanexus/dx-cwl.git
-    bcbiovm_python dx-cwl/dx-cwl compile-workflow $PNAME-workflow/main-$PNAME.cwl --project $DX_PROJECT_ID --token $DX_AUTH_TOKEN --assets ${BCBIO_CONTAINER} --rootdir $output_folder
 
-    generated_cwl=$(dx upload $PNAME-generated-cwl.tgz --brief)
-    dx-jobutil-add-output generated_cwl "$generated_cwl" --class=file
-    
-    dx rm -a $DX_PROJECT_ID:/${output_folder}/main-$PNAME-samples.json || true
-    dx upload --verbose --wait -p --path "$DX_PROJECT_ID:/$output_folder/main-$PNAME-samples.json" $PNAME-workflow/main-$PNAME-samples.json
-    # Wait for upload to complete and the file to be available
-    sleep 5
-
-    bcbiovm_python dx-cwl/dx-cwl run-workflow /$output_folder/main-$PNAME/main-$PNAME /$output_folder/main-$PNAME-samples.json --project $DX_PROJECT_ID --token $DX_AUTH_TOKEN --rootdir $output_folder
+    if [ -n "$reuse_workflow" ]; then
+        WF_ID_OR_NAME=$reuse_workflow
+    else
+        WF_ID_OR_NAME="/$output_folder/main-$PNAME/main-$PNAME"
+        bcbiovm_python dx-cwl/dx-cwl compile-workflow $PNAME-workflow/main-$PNAME.cwl --project $DX_PROJECT_ID --token $DX_AUTH_TOKEN ${BCBIO_ASSETS} --rootdir $output_folder
+        generated_cwl=$(dx upload $PNAME-generated-cwl.tgz --brief)
+        dx-jobutil-add-output generated_cwl "$generated_cwl" --class=file
+        
+        dx rm -a $DX_PROJECT_ID:/${output_folder}/main-$PNAME-samples.json || true
+        dx upload --verbose --wait -p --path "$DX_PROJECT_ID:/$output_folder/main-$PNAME-samples.json" $PNAME-workflow/main-$PNAME-samples.json
+        # Wait for upload to complete and the file to be available
+        sleep 5
+    fi
+   
+    bcbiovm_python dx-cwl/dx-cwl run-workflow $WF_ID_OR_NAME /$output_folder/main-$PNAME-samples.json --project $DX_PROJECT_ID --token $DX_AUTH_TOKEN --rootdir $output_folder
 }
